+++ buildin/buildin/tasks/views.py
@@ -2,7 +2,7 @@ from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
 
 from buildin.repository.account_repository import get_user_full_name
-from buildin.repository.project_repository import get_project_by_slug
+from buildin.repository.project_repository import get_project_by_slug, get_project_participants
 from buildin.repository.task_repository import get_task_by_slug
 
 from buildin.tasks.forms import CreateTaskForm, EditTaskForm, DeleteTaskForm
@@ -32,6 +32,9 @@ def task_create(request, build_slug):
 def task_edit(request, build_slug, task_slug):
     project = get_project_by_slug(build_slug)
     task = get_task_by_slug(task_slug)
+    participants = get_project_participants(project)
+    if not request.user == project.owner and request.user not in participants:
+        return render(request, '403.html')
     if request.method == 'GET':
         form = EditTaskForm(instance=task, project=project)
     else:
@@ -52,6 +55,9 @@ def task_edit(request, build_slug, task_slug):
 def task_delete(request, build_slug, task_slug):
     project = get_project_by_slug(build_slug)
     task = get_task_by_slug(task_slug)
+    participants = get_project_participants(project)
+    if not request.user == project.owner and request.user not in participants:
+        return render(request, '403.html')
     if request.method == 'GET':
         form = DeleteTaskForm(instance=task)
     else:
